LayoutTests/accessibility/ios-simulator/input-type-time.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input id="timeinput" type="time" aria-label="(Between: 07:00 AM-09:00 AM)" min="07:00" max="09:00" step="300" placeholder="00:00 AM" value="08:00">
<input id="timeinput2" type="time" min="07:00" max="09:00" step="300" value="08:00">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that input type: time is accessible on iOS.");

    if (window.accessibilityController) {

        var timeinput = accessibilityController.accessibleElementById("timeinput");
        var timeinput2 = accessibilityController.accessibleElementById("timeinput2");
        shouldBe("timeinput.description", "'AXLabel: (Between: 07:00 AM-09:00 AM)'");
        shouldBeFalse("timeinput.isIgnored");
        
        // Test value is exposed in AXValue.
        shouldBe("timeinput.stringValue", "'AXValue: 8:00 AM'");
        shouldBe("timeinput2.description", "'AXLabel: '");
        shouldBe("timeinput2.stringValue", "'AXValue: 8:00 AM'");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
